Refactor cross-entity relationship definitions and remove unused components
