--- plugin/sdk/pagoda_plugin_sdk/cross_entity/__init__.py
@@ -1,15 +1,0 @@
-"""Cross-entity relationship definitions for pagoda-plugin-sdk."""
-
-from pagoda_plugin_sdk.cross_entity.relationships import (
-    CircularReferenceError,
-    EntityRelationship,
-    RelationshipRegistry,
-    RelationType,
-)
-
-__all__ = [
-    "EntityRelationship",
-    "RelationType",
-    "RelationshipRegistry",
-    "CircularReferenceError",
-]
